+++ src/include/types.h
@@ -5357,7 +5357,7 @@ struct guncmd {
 	u8 type;
 	u8 unk01;
 	u16 unk02;
-	s32 unk04;
+	uintptr_t unk04;
 };
 
 struct pakthing {
